+++ addon/components/lt-foot.js
@@ -31,6 +31,6 @@ import TableHeaderMixin from '../mixins/table-header';
 
 export default Ember.Component.extend(TableHeaderMixin, {
   layout,
-  classNames: ['lt-foot'],
+  classNames: ['lt-foot-wrap'],
   table: null
 });