+++ addon/mixins/table-scroll.js
@@ -3,6 +3,7 @@ import Ember from 'ember';
 const {
   $,
   run,
+  isNone,
   computed
 } = Ember;
 
@@ -47,6 +48,7 @@ export default Ember.Mixin.create({
   didInsertElement() {
     this._super(...arguments);
     if(this.get('_shouldSetupScroll')) {
+      run.scheduleOnce('render', this, this._setupScrollContainer);
       run.scheduleOnce('afterRender', this, this._setupScrollEvents);
     }
   },
@@ -55,6 +57,28 @@ export default Ember.Mixin.create({
     this._super(...arguments);
     if(this.get('_shouldSetupScroll')) {
       this._teardownScrollEvents();
+    }
+  },
+
+  _setupScrollContainer() {
+    const fixedHeader = this.$('.lt-head-wrap thead.lt-head');
+    const fixedFooter = this.$('.lt-foot-wrap tfoot.lt-foot');
+
+    /**
+     * If there is a fixed header and footer, that means that the scroll container
+     * must be the lt-body-wrap with the fixed height and overflow.
+     */
+    if(fixedHeader.length > 0 || fixedFooter.length > 0) {
+      const container = `#${this.get('elementId')} .lt-body-wrap`;
+
+      // Only apply if custom container not specified
+      if(isNone(this.get('attrs.scrollContainer'))) {
+        this.set('scrollContainer', container);
+      }
+
+      if(isNone(this.get('attrs.touchMoveContainer'))) {
+        this.set('touchMoveContainer', container);
+      }
     }
   },